Fix NP GB Memory documentation
It seems like I made some mistakes while testing the size of the map
region. It is actually 256 bytes, not 128. I noticed this when I was
testing it on the flash chip in the Net de Get cartridges.

Also I didn't include the "unprotect sector 0" command. I unlocked
sector 0 on the two cartridges I used for testing, and then I thought
the command isn't necessary, because everything worked without it. It
didn't cross my mind that this command could be persistent, even after
power cycling the cartridge. Now I also found the counterpart command
for re-enabing the protection of sector 0.